MyApps-13: Move track implementation and api to submodules.
