--- app/src/main/java/xyz/torquato/myapps/data/track/TrackRepository.kt
@@ -1,4 +1,0 @@
-package xyz.torquato.myapps.data.track
-
-class TrackRepository {
-}
