+++ data/impl/src/main/java/xyz/torquato/myapps/data/impl/track/external/Track.kt
@@ -1,4 +1,4 @@
-package xyz.torquato.myapps.data.track.datasource
+package xyz.torquato.myapps.data.impl.track.external
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
