+++ data/impl/src/main/java/xyz/torquato/myapps/data/impl/track/external/TrackDatabase.kt
@@ -1,4 +1,4 @@
-package xyz.torquato.myapps.data.track.datasource
+package xyz.torquato.myapps.data.impl.track.external
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
