+++ domain/api/src/main/java/xyz/torquato/myaps/domain/api/track/ITrackRepository.kt
@@ -1,0 +1,6 @@
+package xyz.torquato.myaps.domain.api.track
+
+interface ITrackRepository {
+
+    fun insert(id: Int, title: String?)
+}
